+++ flaskdoc/swagger/__init__.py
@@ -1,12 +1,4 @@
-from flaskdoc.swagger.component import Component
-from flaskdoc.swagger.core import SwaggerDict, SwaggerBase, OpenApi
-from flaskdoc.swagger.info.contact import Contact
-from flaskdoc.swagger.info.info import Info
-from flaskdoc.swagger.info.license import License
-from flaskdoc.swagger.parameters import CookieParameter, HeaderParameter, Parameter, ParameterLocation, PathParameter, \
-    Style, QueryParameter
-from flaskdoc.swagger.path.operations import HttpMethod, Operation
-from flaskdoc.swagger.path.operations import GET, HEAD, OPTIONS, PATCH, POST, PUT, TRACE
-from flaskdoc.swagger.path.paths import PathItem, Paths
-from flaskdoc.swagger.server import Server, ServerVariable
-from flaskdoc.swagger.tag import Tag
+from swagger.models import License, Contact, Info, Server, ServerVariable, Component, Paths, PathItem, Operation, GET, \
+    POST, PUT, HEAD, OPTIONS, PATCH, TRACE, HttpMethod, ParameterLocation, Style, Parameter, PathParameter, \
+    QueryParameter, HeaderParameter, CookieParameter, Tag
+from swagger.core import OpenApi
