--- flaskdoc/swagger/info/__init__.py
@@ -1,3 +1,0 @@
-from flaskdoc.swagger.info.contact import Contact
-from flaskdoc.swagger.info.info import Info
-from flaskdoc.swagger.info.license import License
