+++ tests/test_models.py
@@ -1,16 +1,15 @@
-import swagger
-from swagger import License, Contact
+from flaskdoc.swagger.models import License, Contact
 
 
 def test_license_model():
     li = License(name="Smith L.")
-    d = li.as_dict()
+    d = li.dict()
 
     assert "name" in d
     assert "description" not in d
 
     li = License(name="Smith L.", url="https://www.example.com/license")
-    d = li.as_dict()
+    d = li.dict()
 
     assert d["name"] == "Smith L."
     assert d["url"] == "https://www.example.com/license"
@@ -21,13 +20,13 @@ def test_contact_model():
     # assert
     c1 = Contact()
 
-    d = c1.as_dict()
+    d = c1.dict()
     assert d == {}
 
     c1.name = "Test Smith"
     c1.email = "smith@gmail.com"
 
-    d = c1.as_dict()
+    d = c1.dict()
 
     assert d["name"] == "Test Smith"
     assert d["email"] == "smith@gmail.com"
@@ -39,7 +38,7 @@ def test_server_variable():
 
     assert sv == sv2
 
-    d = sv.as_dict()
+    d = sv.dict()
     assert d["default"] == "1.0"
     assert "enum" not in d
 
@@ -48,7 +47,7 @@ def test_server():
     s1 = swagger.Server(url="https://api.dund.com/{version}", description="Service Endpoint")
     s1.add_variable("version", swagger.ServerVariable("1.1", enum_values=["1.0", "1.1"]))
 
-    d = s1.as_dict()
+    d = s1.dict()
     assert d["variables"]
 
 
@@ -70,6 +69,6 @@ def test_path_items():
         post=post_op
     )
 
-    swag = path_item.as_dict()
+    swag = path_item.dict()
     assert swag["parameters"][0]["name"] == "v1"
     assert swag["get"]["parameters"][0]["name"] == "p1"
